Add conditional security exception handler
