+++ shared-lib/shared-starters/starter-core/src/main/java/com/shared/starter_core/web/SecurityExceptionHandler.java
@@ -1,0 +1,27 @@
+package com.shared.starter_core.web;
+
+import com.common.dto.BaseResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.context.request.WebRequest;
+
+/**
+ * Handles security-related exceptions if Spring Security is present.
+ */
+@Slf4j
+@RestControllerAdvice
+@ConditionalOnClass(AccessDeniedException.class)
+public class SecurityExceptionHandler {
+
+    @ExceptionHandler(AccessDeniedException.class)
+    public ResponseEntity<BaseResponse<?>> handleAccessDenied(AccessDeniedException ex, WebRequest request) {
+        log.warn("Access denied: {}", ex.getMessage());
+        return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                .body(BaseResponse.error("ERR_ACCESS_DENIED", "Access denied"));
+    }
+}
